pyproject.toml
[project]
name = "book-service"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "aiosmtplib==4.0.1",
    "alembic==1.13.3",
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "async-timeout==5.0.1",
    "asyncio==4.0.0",
    "billiard==4.2.1",
    "celery==5.5.3",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "dnspython==2.7.0",
    "email-validator==2.2.0",
    "exceptiongroup==1.3.0",
    "fastapi==0.116.1",
    "flake8==7.3.0",
    "greenlet==3.2.3",
    "h11==0.16.0",
    "httpx==0.27.2",
    "idna==3.10",
    "iniconfig==2.1.0",
    "kombu==5.5.4",
    "mccabe==0.7.0",
    "packaging==25.0",
    "passlib==1.7.4",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.51",
    "psycopg2-binary==2.9.6",
    "pycodestyle==2.14.0",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydantic-settings==2.10.1",
    "pyflakes==3.4.0",
    "pygments==2.19.2",
    "pytest==8.4.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.0.1",
    "python-jose==3.3.0",
    "redis==5.2.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy==2.0.42",
    "starlette==0.47.2",
    "tomli==2.2.1",
    "typing-extensions==4.14.1",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "uvicorn==0.35.0",
    "vine==5.1.0",
    "wcwidth==0.2.13",
]
